Rename to teamcity-oidc-plugin; adopt com.octopus.teamcity.oidc namespace (#3)
