+++ oidc-plugin-server/pom.xml
@@ -3,11 +3,11 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <artifactId>jwt-plugin</artifactId>
-        <groupId>de.ndr.teamcity</groupId>
+        <artifactId>teamcity-oidc-plugin</artifactId>
+        <groupId>com.octopus.teamcity</groupId>
         <version>1.0-SNAPSHOT</version>
     </parent>
-    <artifactId>jwt-plugin-server</artifactId>
+    <artifactId>oidc-plugin-server</artifactId>
     <packaging>jar</packaging>
     <dependencies>
