+++ oidc-plugin-server/src/main/java/com/octopus/teamcity/oidc/JwtBuildFeature.java
@@ -1,4 +1,4 @@
-package de.ndr.teamcity;
+package com.octopus.teamcity.oidc;
 
 import com.nimbusds.jose.JWSAlgorithm;
 import com.nimbusds.jose.jwk.JWK;
