+++ oidc-plugin-server/src/main/resources/buildServerResources/jwtBuildFeatureSettings.jsp
