+++ pom.xml
@@ -2,8 +2,8 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>de.ndr.teamcity</groupId>
-    <artifactId>jwt-plugin</artifactId>
+    <groupId>com.octopus.teamcity</groupId>
+    <artifactId>teamcity-oidc-plugin</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>pom</packaging>
     <scm>
@@ -92,7 +92,7 @@
         </pluginManagement>
     </build>
     <modules>
-        <module>jwt-plugin-server</module>
+        <module>oidc-plugin-server</module>
         <module>build</module>
     </modules>
 </project>
