+++ teamcity-plugin.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <teamcity-plugin xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:schemas-jetbrains-com:teamcity-plugin-v1-xml">
   <info>
-    <name>jwt-plugin</name>
-    <display-name>jwt-plugin</display-name>
+    <name>teamcity-oidc-plugin</name>
+    <display-name>OIDC Identity Token</display-name>
     <version>@Version@</version>
     <description>Plugin description</description>
     <download-url>Plugin download URL</download-url>
